requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
celery-with-redis==3.0
Django==2.0.2
django-celery==3.2.2
kombu==3.0.37
Pillow==5.0.0
PyMySQL==0.8.0
pytz==2018.3
redis==2.10.6
vine==1.1.4
